Moved vision helpers into utility package
