+++ src/main/java/frc/robot/subsystems/GCPoseEstimator.java
@@ -21,7 +21,7 @@ import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Constants.DriveConstants;
-import frc.robot.LimelightHelpers;
+import frc.robot.utility.LimelightHelpers;
 
 //TODO: Remove this suppress warnings after finishing
 @SuppressWarnings("unused")
