+++ src/main/java/frc/robot/utility/GCLimelight.java
@@ -1,12 +1,12 @@
 
-package frc.robot;
+package frc.robot.utility;
 
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.estimator.PoseEstimator;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
-import frc.robot.LimelightHelpers.LimelightResults;
 import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.utility.LimelightHelpers.LimelightResults;
 
 
 public class GCLimelight {
