+++ src/main/java/frc/robot/utility/GCPhotonVision.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
- package frc.robot;
+ package frc.robot.utility;
 
  import static frc.robot.Constants.VisionConstants.*;
  
@@ -37,7 +37,8 @@ import edu.wpi.first.math.numbers.N1;
  import edu.wpi.first.math.numbers.N3;
  import edu.wpi.first.math.util.Units;
  import edu.wpi.first.wpilibj.smartdashboard.Field2d;
- import frc.robot.Constants.VisionConstants;
+import frc.robot.Robot;
+import frc.robot.Constants.VisionConstants;
  
  import java.util.List;
  import java.util.Optional;
